+++ ThirdPartyAdapters/pangle/pangle/src/main/java/com/google/ads/mediation/pangle/PangleMediationAdapter.java
@@ -12,6 +12,7 @@ import com.bytedance.sdk.openadsdk.TTAdManager;
 import com.bytedance.sdk.openadsdk.TTAdSdk;
 import com.google.ads.mediation.pangle.rtb.PangleRtbBannerAd;
 import com.google.ads.mediation.pangle.rtb.PangleRtbInterstitialAd;
+import com.google.ads.mediation.pangle.rtb.PangleRtbNativeAd;
 import com.google.ads.mediation.pangle.rtb.PangleRtbRewardedAd;
 import com.google.android.gms.ads.AdError;
 import com.google.android.gms.ads.MobileAds;
@@ -45,6 +46,7 @@ public class PangleMediationAdapter extends RtbAdapter {
   private PangleRtbBannerAd bannerAd;
   private PangleRtbInterstitialAd interstitialAd;
   private PangleRtbRewardedAd rewardedAd;
+  private PangleRtbNativeAd nativeAd;
   private static int coppa = -1;
   private static int gdpr = -1;
   private static int ccpa = -1;
@@ -163,9 +165,8 @@ public class PangleMediationAdapter extends RtbAdapter {
   @Override
   public void loadRtbInterstitialAd(
       @NonNull MediationInterstitialAdConfiguration adConfiguration,
-      @NonNull
-          MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback>
-              callback) {
+      @NonNull MediationAdLoadCallback<MediationInterstitialAd, MediationInterstitialAdCallback>
+          callback) {
     interstitialAd = new PangleRtbInterstitialAd(adConfiguration, callback);
     interstitialAd.render();
   }
@@ -174,10 +175,8 @@ public class PangleMediationAdapter extends RtbAdapter {
   public void loadRtbNativeAd(
       @NonNull MediationNativeAdConfiguration adConfiguration,
       @NonNull MediationAdLoadCallback<UnifiedNativeAdMapper, MediationNativeAdCallback> callback) {
-    AdError error =
-        new AdError(0, "Native ad is not supported in Pangle.", PangleConstants.ERROR_DOMAIN);
-    Log.w(TAG, error.toString());
-    callback.onFailure(error);
+    nativeAd = new PangleRtbNativeAd(adConfiguration, callback);
+    nativeAd.render();
   }
 
   @Override
@@ -196,10 +195,10 @@ public class PangleMediationAdapter extends RtbAdapter {
    * Set the COPPA setting in Pangle SDK.
    *
    * @param coppa an {@code Integer} value that indicates whether the app should be treated as
-   *     child-directed for purposes of the COPPA. {@link
-   *     RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE} means true. {@link
-   *     RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_FALSE} means false. {@link
-   *     RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED} means unspecified.
+   *              child-directed for purposes of the COPPA. {@link RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_TRUE}
+   *              means true. {@link RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_FALSE}
+   *              means false. {@link RequestConfiguration#TAG_FOR_CHILD_DIRECTED_TREATMENT_UNSPECIFIED}
+   *              means unspecified.
    */
   public static void setCoppa(@TagForChildDirectedTreatment int coppa) {
     switch (coppa) {
@@ -228,9 +227,9 @@ public class PangleMediationAdapter extends RtbAdapter {
    * Set the GDPR setting in Pangle SDK.
    *
    * @param gdpr an {@code Integer} value that indicates whether the user consents the use of
-   *     personal data to serve ads under GDPR. {@code 0} means the user consents. {@code 1} means
-   *     the user does not consent. {@code -1} means the user hasn't specified. Any value outside of
-   *     -1, 0, or 1 will result in this method being a no-op.
+   *             personal data to serve ads under GDPR. {@code 0} means the user consents. {@code 1}
+   *             means the user does not consent. {@code -1} means the user hasn't specified. Any
+   *             value outside of -1, 0, or 1 will result in this method being a no-op.
    */
   public static void setGdpr(int gdpr) {
     if (gdpr != 0 && gdpr != 1 && gdpr != -1) {
@@ -248,9 +247,9 @@ public class PangleMediationAdapter extends RtbAdapter {
    * Set the CCPA setting in Pangle SDK.
    *
    * @param ccpa an {@code Integer} value that indicates whether the user opts in of the "sale" of
-   *     the "personal information" under CCPA. {@code 0} means the user opts in. {@code 1} means
-   *     the user opts out. {@code -1} means the user hasn't specified. Any value outside of -1, 0,
-   *     or 1 will result in this method being a no-op.
+   *             the "personal information" under CCPA. {@code 0} means the user opts in. {@code 1}
+   *             means the user opts out. {@code -1} means the user hasn't specified. Any value
+   *             outside of -1, 0, or 1 will result in this method being a no-op.
    */
   public static void setCcpa(int ccpa) {
     if (ccpa != 0 && ccpa != 1 && ccpa != -1) {